Moved things around after the merge.
