+++ TeamCode/src/main/java/com/edinaftc/library/subsystems/MecanumDrive.java
@@ -1,7 +1,5 @@
-package com.edinaftc.subsystems;
+package com.edinaftc.library.subsystems;
 
-
-import android.graphics.PorterDuff;
 
 import com.edinaftc.library.Vector2d;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -9,8 +7,6 @@ import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.PIDCoefficients;
-import com.qualcomm.robotcore.hardware.PIDFCoefficients;
-import com.qualcomm.robotcore.util.Hardware;
 
 import java.util.Arrays;
 import java.util.Collections;
